Template test

+++ src/baofit/create_samples_files.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 BOX="1"
 WORKDIR=/hpcstorage/dforero/projects/baosystematics/patchy_results/box${BOX}
+OVERWRITE=0
 for case_ in 1 2 3
 do
 for space in real redshift
@@ -10,21 +11,27 @@ do
 
 files=$WORKDIR/$space/nosyst/baofit/${tracer}/*mystats*
 ofile=$WORKDIR/$space/nosyst/baofit/${tracer}/alpha_samples.dat
+if [[ ! -e $ofile ]] || [[ ${OVERWRITE} -eq 1 ]]; then
 rm -v $ofile
 cat $files >> $ofile
+fi
 wc -l $ofile
 for map in smooth #noise
 do
 files=$WORKDIR/$space/$map/parabola_0.8/baofit/${tracer}/*mystats*
 ofile=$WORKDIR/$space/$map/parabola_0.8/baofit/${tracer}/alpha_samples.dat
+if [[ ! -e $ofile ]] || [[ ${OVERWRITE} -eq 1 ]]; then
 rm -v $ofile
 cat $files >> $ofile
+fi
 wc -l $ofile
 done
 files=$WORKDIR/$space/radialgauss/baofit/${tracer}/*mystats*
 ofile=$WORKDIR/$space/radialgauss/baofit/${tracer}/alpha_samples.dat
+if [[ ! -e $ofile ]] || [[ ${OVERWRITE} -eq 1 ]]; then
 rm -v $ofile
 cat $files >> $ofile
+fi
 wc -l $ofile
 done
 done
